Move TFC to /module/ folder
