+++ src/main/java/gregification/modules/tfc/ITFCFluidRegistrationStatus.java
@@ -1,4 +1,4 @@
-package gregification.tfc;
+package gregification.modules.tfc;
 
 public interface ITFCFluidRegistrationStatus {
